# apidog/elloot-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elloot API",
    "description": "Marketplace API (escrow). Import this file in Apidog: Import → OpenAPI.\n\nAuth: use `Authorization: Bearer <accessToken>` from register/login.\nSandbox flow: create order → checkout → POST /payments/sandbox/confirm → deliver → confirm.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local (.env.local PORT=5000)"
    },
    {
      "url": "http://localhost:3333",
      "description": "Local default (.env.example)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT accessToken from /api/auth/register or /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "UNAUTHORIZED" },
              "message": { "type": "string", "example": "Unauthorized" }
            }
          }
        }
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "role": {
            "type": "string",
            "enum": ["BUYER", "SELLER", "ADMIN"]
          },
          "kycStatus": {
            "type": "string",
            "enum": ["NONE", "PENDING", "APPROVED", "REJECTED"]
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["user", "accessToken"],
        "properties": {
          "user": { "$ref": "#/components/schemas/UserPublic" },
          "accessToken": { "type": "string" }
        }
      },
      "RegisterBody": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8, "maxLength": 72 },
          "name": { "type": "string", "minLength": 2, "maxLength": 80 }
        }
      },
      "LoginBody": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 1 }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register with email/password",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterBody" },
              "example": {
                "email": "seller@elloot.test",
                "password": "password123",
                "name": "Seller"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "409": {
            "description": "Email taken",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email/password",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginBody" },
              "example": {
                "email": "seller@elloot.test",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user",
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user profile" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Auth"],
        "summary": "Update profile (name, avatarUrl, pixKey)",
        "operationId": "patchMe",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "nullable": true },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "pixKey": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated profile" } }
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Google OAuth callback",
        "operationId": "googleCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "state",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Use `json` to return tokens as JSON instead of frontend redirect",
            "schema": { "type": "string", "enum": ["json"] }
          }
        ],
        "responses": {
          "200": { "description": "JSON auth result when format=json" },
          "302": { "description": "Redirect to frontend with token" }
        }
      }
    },
...
    "/api/auth/discord/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Discord OAuth callback",
        "operationId": "discordCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "state",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["json"] }
          }
        ],
        "responses": {
          "200": { "description": "JSON auth result when format=json" },
          "302": { "description": "Redirect to frontend with token" }
        }
      }
    },
...
  }
}